+++ app/util/FileUtils.scala
@@ -1,6 +1,6 @@
 package util
 
-import java.io.FileInputStream
+import java.io.{File => JFile}
 import java.net.URL
 import java.util.Date
 
@@ -350,6 +350,31 @@ object FileUtils {
     }
 
     uploadedFiles.toList
+  }
+
+  // process a file from a bagit zipfile upload
+  def processBagFile(file: java.io.File, filename: String, user: User, dataset: Dataset, folderId: Option[String]): Option[File] = {
+    // Save file bytes
+    ByteStorageService.save(file, "uploads") match {
+      case Some((loader_id, loader, length)) => {
+        val fileobj = File(UUID.generate, loader_id, filename, filename, user, new Date(), getContentType(filename, None),
+          file.length, loader, licenseData=License.fromAppConfig, stats=new Statistics(), status=FileStatus.PROCESSED.toString)
+        files.save(fileobj)
+        appConfig.incrementCount('files, 1)
+        appConfig.incrementCount('bytes, file.length)
+
+        // Add to dataset/folder if given
+        folderId match {
+          case Some(fid) => folders.addFile(UUID(fid), fileobj.id)
+          case None => datasets.addFile(dataset.id, fileobj)
+        }
+        Some(fileobj)
+      }
+      case None => {
+        Logger.error("Unable to save file bytes to Clowder.")
+        None
+      }
+    }
   }
 
   // process a single uploaded file