+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Log an event each time that a user archives or unarchives a file.
 - Endpoint `/api/datasets/createfrombag` to ingest datasets in BagIt format. Includes basic dataset metadata, files,
   folders and technical metadata. Downloading datasets now includes extra Datacite and Clowder metadata.
+- Endpoint /api/files/bulkRemove to delete multiple files in one call. [#12](https://github.com/clowder-framework/clowder/issues/12)
 
 ## 1.16.0 - 2021-03-31